Make selection bg color opaque
This still needs tweaking, but this is part of the fundamental problem
of not being able to see what is selected when FIND ALL or FIND is in
use. My pink selection color with low opacity was getting obliterated
by find highlight colors even though they also had low opacity... the
color blending is hard to get right, which is I guess why almost all
themes have this problem.